% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_bio_c.R
\docType{data}
\name{cr_bio_c}
\alias{cr_bio_c}
\title{Bioclimatic variables of continental Costa Rica}
\format{
A stars object with the bioclimatic variables of continental Costa Rica (EPSG:4326)
}
\source{
The data was obtained using the R package geodata. The data source is WorldClim version 2.1 \url{https://www.worldclim.org/}
}
\usage{
cr_bio_c
}
\description{
Bioclimatic variables of continental Costa Rica, resolution 1km (0.5 minutes of a degree)
}
\keyword{datasets}
