% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_cantons.R
\docType{data}
\name{cr_cantons}
\alias{cr_cantons}
\title{Cantons of Costa Rica}
\format{
sf object of cantons of Costa Rica (EPSG:4326)
}
\source{
\url{https://gadm.org/index.html}
}
\usage{
cr_cantons
}
\description{
Cantons of Costa Rica
}
\keyword{datasets}
