% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_cantons_c.R
\docType{data}
\name{cr_cantons_c}
\alias{cr_cantons_c}
\title{Cantons of Costa Rica (continent)}
\format{
sf object of cantons of Costa Rica (EPSG:4326)
}
\source{
\url{https://gadm.org/index.html}
}
\usage{
cr_cantons_c
}
\description{
Cantons of Costa Rica (without islands)
}
\keyword{datasets}
