% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_districts.R
\docType{data}
\name{cr_districts}
\alias{cr_districts}
\title{Districts of Costa Rica}
\format{
sf object of districts of Costa Rica (EPSG:4326)
}
\source{
\url{https://gadm.org/index.html}
}
\usage{
cr_districts
}
\description{
Districts of Costa Rica
}
\keyword{datasets}
