% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_districts_c.R
\docType{data}
\name{cr_districts_c}
\alias{cr_districts_c}
\title{Districts of Costa Rica (continent)}
\format{
sf object of districts of Costa Rica (EPSG:4326)
}
\source{
\url{https://gadm.org/index.html}
}
\usage{
cr_districts_c
}
\description{
Districts of Costa Rica (without islands)
}
\keyword{datasets}
