% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_elevation.R
\docType{data}
\name{cr_elevation}
\alias{cr_elevation}
\title{Elevation data of Costa Rica}
\format{
A stars object with elevation data of Costa Rica (EPSG:4326)
}
\source{
The data was obtained using the R package geodata The main data source is Shuttle Radar Topography Mission (SRTM)
}
\usage{
cr_elevation
}
\description{
Elevation data of Costa Rica, resolution 1km (30 arc seconds)
}
\keyword{datasets}
