% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_elevation_c.R
\docType{data}
\name{cr_elevation_c}
\alias{cr_elevation_c}
\title{Elevation data of continenetal Costa Rica}
\format{
A stars object with elevation data of continental Costa Rica (EPSG:4326)
}
\source{
The data was obtained using the R package geodata The main data source is Shuttle Radar Topography Mission (SRTM)
}
\usage{
cr_elevation_c
}
\description{
Elevation data of continental Costa Rica, resolution 1km (30 arc seconds)
}
\keyword{datasets}
