% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_grids.R
\docType{data}
\name{cr_grid_8km}
\alias{cr_grid_8km}
\alias{cr_grid_4km}
\alias{cr_grid_2km}
\alias{cr_grid_1km}
\title{Hierachical spatial grid system for Costa Rica}
\format{
A sf object with a spatial grid for Costa Rica (EPSG:5367)
}
\source{
The map of Costa Rica was taken from \url{https://gadm.org/index.html}
}
\usage{
cr_grid_8km
}
\description{
Spatial square grids of 8km, 4km, 2km, and 1km for Costa Rica
}
\keyword{datasets}
