% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_osm_highways.R
\docType{data}
\name{cr_osm_highways}
\alias{cr_osm_highways}
\title{Highways of Costa Rica}
\format{
An sf object with highways of Costa Rica (EPSG:4326)
}
\source{
\url{https://www.openstreetmap.org}
}
\usage{
cr_osm_highways
}
\description{
OpenStreetMap highways of Costa Rica
}
\keyword{datasets}
