% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_osm_places.R
\docType{data}
\name{cr_osm_places}
\alias{cr_osm_places}
\title{Places of Costa Rica}
\format{
An sf object with places of Costa Rica (EPSG:4326)
}
\source{
\url{https://www.openstreetmap.org}
}
\usage{
cr_osm_places
}
\description{
OpenStreetMap places of Costa Rica
}
\keyword{datasets}
