% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_osm_railway.R
\docType{data}
\name{cr_osm_railway}
\alias{cr_osm_railway}
\title{Railway of Costa Rica}
\format{
An sf object with the railway of Costa Rica (EPSG:4326)
}
\source{
\url{https://www.openstreetmap.org}
}
\usage{
cr_osm_railway
}
\description{
OpenStreetMap railway of Costa Rica
}
\keyword{datasets}
