% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_outline.R
\docType{data}
\name{cr_outline}
\alias{cr_outline}
\title{Outline of Costa Rica}
\format{
sf object of the outline of Costa Rica (EPSG:4326)
}
\source{
\url{https://gadm.org/index.html}
}
\usage{
cr_outline
}
\description{
Outline of Costa Rica
}
\keyword{datasets}
