% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_outline_c.R
\docType{data}
\name{cr_outline_c}
\alias{cr_outline_c}
\title{Outline of continental Costa Rica}
\format{
sf object of the outline of Costa Rica (EPSG:4326)
}
\source{
\url{https://gadm.org/index.html}
}
\usage{
cr_outline_c
}
\description{
Outline of continental Costa Rica (without islands)
}
\keyword{datasets}
