% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_pop_density_2020_1km.R
\docType{data}
\name{cr_pop_density_2020_1km}
\alias{cr_pop_density_2020_1km}
\title{Population density of Costa Rica 2020}
\format{
stars object of the population density of Costa Rica (EPSG:4326)
}
\source{
\url{https://gadm.org/index.html}
}
\usage{
cr_pop_density_2020_1km
}
\description{
Population density of Costa Rica
}
\keyword{datasets}
