% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_prec_c.R
\docType{data}
\name{cr_prec_c}
\alias{cr_prec_c}
\title{Total precipitation data of continental Costa Rica}
\format{
A stars object with the total precipitation data of continental Costa Rica (EPSG:4326)
}
\source{
The data was obtained using the R package geodata. The data source is WorldClim version 2.1 \url{https://www.worldclim.org/}
}
\usage{
cr_prec_c
}
\description{
Total precipitation data of continental Costa Rica (mm), resolution 1km (0.5 minutes of a degree)
}
\keyword{datasets}
