% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_provinces.R
\docType{data}
\name{cr_provinces}
\alias{cr_provinces}
\title{Provinces of Costa Rica}
\format{
sf object of provinces of Costa Rica (EPSG:4326)
}
\source{
\url{https://gadm.org/index.html}
}
\usage{
cr_provinces
}
\description{
Provinces of Costa Rica
}
\keyword{datasets}
