% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_provinces_c.R
\docType{data}
\name{cr_provinces_c}
\alias{cr_provinces_c}
\title{Provinces of Costa Rica (continent)}
\format{
sf object of provinces of Costa Rica (EPSG:4326)
}
\source{
\url{https://gadm.org/index.html}
}
\usage{
cr_provinces_c
}
\description{
Provinces of Costa Rica (without islands)
}
\keyword{datasets}
