% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_tmax.R
\docType{data}
\name{cr_tmax}
\alias{cr_tmax}
\title{Maximum temperature data of Costa Rica}
\format{
A stars object with the maximum temperature data of Costa Rica (EPSG:4326)
}
\source{
The data was obtained using the R package geodata. The data source is WorldClim version 2.1 \url{https://www.worldclim.org/}
}
\usage{
cr_tmax
}
\description{
Maximum temperature data of Costa Rica (°C), resolution 1km (0.5 minutes of a degree)
}
\keyword{datasets}
