% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_tmin.R
\docType{data}
\name{cr_tmin}
\alias{cr_tmin}
\title{Minimum temperature data of Costa Rica}
\format{
A stars object with the minimum temperature data of Costa Rica (EPSG:4326)
}
\source{
The data was obtained using the R package geodata. The data source is WorldClim version 2.1 \url{https://www.worldclim.org/}
}
\usage{
cr_tmin
}
\description{
Minimum temperature data of Costa Rica (°C), resolution 1km (0.5 minutes of a degree)
}
\keyword{datasets}
