% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_wind_c.R
\docType{data}
\name{cr_wind_c}
\alias{cr_wind_c}
\title{Wind speed data of continental Costa Rica}
\format{
A stars object with wind speed data of continental Costa Rica (EPSG:4326)
}
\source{
The data was obtained using the R package geodata. The data source is WorldClim version 2.1 \url{https://www.worldclim.org/}
}
\usage{
cr_wind_c
}
\description{
Wind speed (2m above the ground) data of continental Costa Rica (m s^-1), resolution 1km (0.5 minutes of a degree)
}
\keyword{datasets}
