% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crgeo-package.R
\docType{package}
\name{crgeo-package}
\alias{crgeo}
\alias{crgeo-package}
\title{crgeo: crgeo: Geospatial data of Costa Rica}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Contains geospatial data of Costa Rica.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ManuelSpinola/crgeo}
  \item Report bugs at \url{https://github.com/ManuelSpinola/crgeo/issues}
}

}
\keyword{internal}
